+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.0.2] - 2018-12-22
+### Changed
+- Merged quick screen switch and quick render menu into one
+
+### Added
+- Pie menu and custom position
+
 ## [0.0.1] - 2018-12-20
 ### New
 - Initial addon setup
@@ -12,4 +19,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 <!--### Official Rigify Info-->
 
+[0.0.2]:https://github.com/schroef/QuickSwitch/releases/tag/v.0.0.2
 [0.0.1]:https://github.com/schroef/QuickSwitch/releases/tag/v.0.0.1